fix non-blockers
