+++ mm2src/mm2_main/tests/mm2_tests/orderbook_sync_tests.rs
@@ -1400,6 +1400,7 @@ fn zhtlc_orders_sync_alice_connected_before_creation() {
         ZOMBIE_TICKER,
         ZOMBIE_ELECTRUMS,
         ZOMBIE_LIGHTWALLETD_URLS,
+        None,
     ));
 
     let set_price_json = json!({
@@ -1462,6 +1463,7 @@ fn zhtlc_orders_sync_alice_connected_after_creation() {
         ZOMBIE_TICKER,
         ZOMBIE_ELECTRUMS,
         ZOMBIE_LIGHTWALLETD_URLS,
+        None,
     ));
 
     let set_price_json = json!({
@@ -1490,6 +1492,7 @@ fn zhtlc_orders_sync_alice_connected_after_creation() {
         ZOMBIE_TICKER,
         ZOMBIE_ELECTRUMS,
         ZOMBIE_LIGHTWALLETD_URLS,
+        None,
     ));
 
     let set_price_json = json!({
